Stop ignoring uses for symbols with mappings
In various situations IWYU will ignore symbols.  E.g. if they are
builtins, or defined in the file where they are used.

This makes sense normally.  However, when a symbol has an explicit
mapping defined, we should not ignore uses of it.

In particular, for example, one might require a particular header to be
included when certain builtins are used, and this change allows that by
defining a mapping for that builtin (whereas before all uses of builtins
were ignored).